fix: Fix conflicts when publishing tags
Fixes a problem when trying to publish tags from a given commit
after a newer commit has been pushed to the currently checked
out branch.

This can happen when the Github action runs on the main branch
and two PRs are merged in quick succession. The first release
will fail to push because there are newer commits on main.