feat: Support pre-release versions in `versionName` resolution (#226)
Ensure the Gradle `Project.versionName` extension returns the latest pre-release (if it exists).

## Context

After #212, it's possible to create a pre-release version (e.g. `1.2.3-alpha.1`). However, this version isn't yet reflected in `Project.versionName` field.

This leads to inconsistency between the version of a commit and the version of the project that's published.

DCMAW-16073