feat: Allow version name to be specified when publishing (#229)
Allow the version name to be specified manually when publishing with the following actions:

- `actions/bundle-release-app`
- `actions/maven-publish-local`
- `actions/maven-publish`

If the version name isn't supplied or an empty value is supplied, the current behaviour is unchanged.

This change makes it possible to specify the exact version name to be published so that publishing workflows can be protected from newly pushed tags changing the latest version.

DCMAW-16073